Add files via upload
This file already achieved the basic logic of this project, it can count how many people in the room, and change LED to corresponding color (number of people & color):
1-9  red;
10-19 yellow;
20-30 green.

However, the current counting logic is: when the first ultrasonic sensor detected a person and the second ultrasonic sensor does not detected a person, it count as someone entered room. If the detection sequence is reversed, it is considered that someone has left the room. This counting logic can applies to the places where entry and exit channels are separate, such as some concert venues. 

I'm working on the iteration of this counting logic in order to apply this project in CE Lab, please look forward to the next version :)