Add files via upload
This file including the method to count how many people in the lab through two ultrasonic  sensors which can be deployed at the same door. The basic logic is "If the first ultrasonic sensor detects a distance less than 5cm, starting a while loop to judge the second ultrasonic sensor", by doing so, we can know the sequence of entering and exiting.